add routes for modeus
